finish the project and all the problems with language
