+++ app/layout.js
@@ -1,23 +1,14 @@
-import { Geist, Geist_Mono } from "next/font/google";
+// app/layout.js
 import "../app/styles/globals.css";
 import I18nProvider from "../components/Providers/I18nProvider";
 import Navbar from "@/components/common/Navbar";
 import Footer from "@/components/common/Footer";
-import { Zain } from 'next/font/google';
+import { detectLanguage } from "@/lib/language-detector";
+import { Zain } from "next/font/google";
 
 const zain = Zain({
-  subsets: ['latin'],
-  weight: [ '200', '300', '400', '700', '800', '900'],
-});
-// import i18n from "../lib/i18n";
-const geistSans = Geist({
-  variable: "--font-geist-sans",
   subsets: ["latin"],
-});
-
-const geistMono = Geist_Mono({
-  variable: "--font-geist-mono",
-  subsets: ["latin"],
+  weight: ["200", "300", "400", "700", "800", "900"],
 });
 
 export const metadata = {
@@ -25,13 +16,20 @@ export const metadata = {
   description: "Generated by create next app",
 };
 
-export default function RootLayout({ children }) {
+export default async function RootLayout({ children }) {
+  // Detect language on server side
+  const lng = await detectLanguage();
+
   return (
-    <html lang="en">
+    <html
+      lang={lng}
+      dir={lng === "ar" ? "rtl" : "ltr"}
+      className={zain.className}
+    >
       <body>
-        <Navbar/>
+        <Navbar />
         <main className="">
-          <I18nProvider>{children}</I18nProvider>
+          <I18nProvider lng={lng}>{children}</I18nProvider>
         </main>
         <Footer />
       </body>
