+++ components/common/VerificationSlider.jsx
@@ -1,3 +1,4 @@
+'use client';
 import React, { useEffect, useState } from "react";
 import ArrowForwardIosIcon from "@mui/icons-material/ArrowForwardIos";
 import ArrowBackIosNewIcon from "@mui/icons-material/ArrowBackIosNew";
