+++ lib/constants.js
@@ -1,0 +1,6 @@
+export const languages = [
+  { code: 'en', name: 'English' },
+  { code: 'ar', name: 'العربية' }
+];
+
+export const defaultLanguage = 'en';
