+++ lib/i18n.js
@@ -1,26 +1,42 @@
-import i18n from 'i18next';
-import { initReactI18next } from 'react-i18next';
-import HttpApi from 'i18next-http-backend';
-import LanguageDetector from 'i18next-browser-languagedetector';
+// lib/i18n.js
+import { createInstance } from "i18next";
+import resourcesToBackend from "i18next-resources-to-backend";
+import { initReactI18next } from "react-i18next/initReactI18next";
 
-i18n
-  .use(initReactI18next)
-  .use(LanguageDetector)
-  .use(HttpApi)
-  .init({
-    fallbackLng: 'en',
-    supportedLngs: ['en', 'ar'],
-     detection: {
-        order: ['cookie','localStorage', 'hash',  'sessionStorage','htmlTag', 'navigator',  'path', 'subdomain'],
-        caches: ['cookie', 'localStorage'],
-     },
-     backend: {
-        loadPath:'/locales/{{lng}}/translation.json',
-     },
-    debug: false,
-    interpolation: {
-      escapeValue: false,
-    },
-  });
+const initI18next = async (lng, ns) => {
+  const i18nInstance = createInstance();
+  await i18nInstance
+    .use(initReactI18next)
+    .use(
+      resourcesToBackend((language, namespace) =>
+        // Ensure this loads correctly on the server
+        import(`../public/locales/${language}/${namespace}.json`)
+      )
+    )
+    .init({
+      lng,
+      fallbackLng: "en",
+      supportedLngs: ["en", "ar"],
+      defaultNS: "translation",
+      fallbackNS: "translation",
+      ns,
+      interpolation: {
+        escapeValue: false,
+      },
+      // Important for server components to pre-load
+      preload: typeof window === 'undefined' ? ["en", "ar"] : [],
+    });
+  return i18nInstance;
+};
 
-export default i18n;
+export async function useTranslation(lng, ns = "translation", options = {}) {
+  const i18nextInstance = await initI18next(lng, ns);
+  return {
+    t: i18nextInstance.getFixedT(
+      lng,
+      Array.isArray(ns) ? ns[0] : ns,
+      options.keyPrefix
+    ),
+    i18n: i18nextInstance,
+  };
+}
