+++ package.json
@@ -21,6 +21,7 @@
     "i18next": "^25.3.2",
     "i18next-browser-languagedetector": "^8.2.0",
     "i18next-http-backend": "^3.0.2",
+    "i18next-resources-to-backend": "^1.2.1",
     "js-cookie": "^3.0.5",
     "lucide-react": "^0.539.0",
     "next": "15.4.6",
